// types/bankAccountDetails.ts
export interface BankAccountDetails {
  id: string;
  accountOwnerName: string;
  hostToHostInterimStatementType: string;
  bankName: string;
  branchName: string;
  bic: string;
  sortCode: string;
  townCity: string;
  country: string;
  accountNumber: string;
  iban: string;
  currency: string;
  accountType: string;
  currencyAndTransactionLimit: string;
  paymentType: string;
}